+++ lib/twilio-ruby/rest/fax/v1/fax.rb
@@ -164,12 +164,10 @@ module Twilio
           #   fax. See our [security](https://www.twilio.com/docs/security) page for
           #   information on how to ensure the request for your media comes from Twilio.
           # @param [fax.Quality] quality The [Fax Quality
-          #   value](https://www.twilio.com/docs/api/fax/rest/faxes#fax-quality-values) that
-          #   describes the fax quality. Can be: `standard`, `fine`, or `superfine` and
-          #   defaults to `fine`.
+          #   value](https://www.twilio.com/docs/api/fax/rest/faxes-resource#fax-quality-values) that describes the fax quality. Can be: `standard`, `fine`, or `superfine` and defaults to `fine`.
           # @param [String] status_callback The URL we should call using the `POST` method
-          #   to send status information to your application when the status of the fax
-          #   changes.
+          #   to send [status
+          #   information](https://www.twilio.com/docs/api/fax/rest/faxes-resource#fax-status-callback) to your application when the status of the fax changes.
           # @param [String] from The number the fax was sent from. Can be the phone number
           #   in [E.164](https://www.twilio.com/docs/glossary/what-e164) format or the SIP
           #   `from` value. The caller ID displayed to the recipient uses this value. If this
@@ -284,8 +282,8 @@ module Twilio
 
           ##
           # Update the FaxInstance
-          # @param [fax.UpdateStatus] status The new status of the resource. Can be only
-          #   `canceled`. This may fail if transmission has already started.
+          # @param [fax.UpdateStatus] status The new
+          #   [status](https://www.twilio.com/docs/api/fax/rest/faxes-resource#fax-status-values) of the resource. Can be only `canceled`. This may fail if transmission has already started.
           # @return [FaxInstance] Updated FaxInstance
           def update(status: :unset)
             data = Twilio::Values.of({'Status' => status, })
@@ -507,8 +505,8 @@ module Twilio
 
           ##
           # Update the FaxInstance
-          # @param [fax.UpdateStatus] status The new status of the resource. Can be only
-          #   `canceled`. This may fail if transmission has already started.
+          # @param [fax.UpdateStatus] status The new
+          #   [status](https://www.twilio.com/docs/api/fax/rest/faxes-resource#fax-status-values) of the resource. Can be only `canceled`. This may fail if transmission has already started.
           # @return [FaxInstance] Updated FaxInstance
           def update(status: :unset)
             context.update(status: status, )