+++ spec/integration/autopilot/v1/assistant_spec.rb
@@ -35,6 +35,8 @@ describe 'Assistant' do
           "friendly_name": "so so friendly",
           "latest_model_build_sid": null,
           "log_queries": true,
+          "development_stage": "in-development",
+          "needs_model_build": false,
           "sid": "UAaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
           "unique_name": "so-so-unique",
           "links": {
@@ -111,6 +113,8 @@ describe 'Assistant' do
                   "friendly_name": "so so friendly",
                   "latest_model_build_sid": null,
                   "log_queries": true,
+                  "development_stage": "in-development",
+                  "needs_model_build": false,
                   "sid": "UAaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
                   "unique_name": "so-so-unique",
                   "links": {
@@ -172,6 +176,8 @@ describe 'Assistant' do
           "friendly_name": "so so friendly",
           "latest_model_build_sid": null,
           "log_queries": true,
+          "development_stage": "in-development",
+          "needs_model_build": false,
           "sid": "UAaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
           "unique_name": "so-so-unique",
           "links": {
@@ -222,6 +228,8 @@ describe 'Assistant' do
           "friendly_name": "so so friendly",
           "latest_model_build_sid": null,
           "log_queries": true,
+          "development_stage": "in-development",
+          "needs_model_build": false,
           "sid": "UAaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
           "unique_name": "so-so-unique",
           "links": {