+++ spec/integration/preview/trusted_comms/current_call_spec.rb
@@ -30,8 +30,8 @@ describe 'CurrentCall' do
       %q[
       {
           "sid": "CAaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
-          "from": "+15000000000",
-          "to": "+573000000000",
+          "from": "+1500123",
+          "to": "+1500456",
           "status": "ringing",
           "reason": "Hello Jhon, your bank appointment has been confirmed.",
           "created_at": "2019-05-01T20:00:00Z",