+++ spec/integration/verify/v2/service/verification_check_spec.rb
@@ -37,9 +37,9 @@ describe 'VerificationCheck' do
           "to": "+14159373912",
           "channel": "sms",
           "status": "approved",
-          "valid": false,
-          "amount": "$29.99",
-          "payee": "Acme",
+          "valid": true,
+          "amount": null,
+          "payee": null,
           "date_created": "2015-07-30T20:00:00Z",
           "date_updated": "2015-07-30T20:00:00Z"
       }